+++ packages/flutter_riverpod/test/providers/change_notifier/change_notifier_provider_test.dart
@@ -6,7 +6,17 @@ import 'package:mockito/mockito.dart';
 import '../../utils.dart';
 
 void main() {
-  test('can read and set current StateNotifier', () async {
+  test('support null ChangeNotifier', () {
+    final container = createContainer();
+    final provider = ChangeNotifierProvider<ValueNotifier<int>?>((ref) => null);
+
+    expect(container.read(provider), null);
+    expect(container.read(provider.notifier), null);
+
+    container.dispose();
+  });
+
+  test('can read and set current ChangeNotifier', () async {
     final container = createContainer();
     final listener = Listener<ValueNotifier<int>>();
     late ChangeNotifierProviderRef<ValueNotifier<int>> ref;